Condense Sidebar options into a directory
