+++ src/components/Sidebar/SidebarControls/index.js
@@ -1,0 +1,3 @@
+import { SidebarControls } from './SidebarControls'
+
+export default SidebarControls;
